Fixed #9156, export data download failed on iOS.
